Implement the `InstructionStream` type
This type represents a sequence of EVM `Opcode`s, and maintains a 1:1
correspondence with the byte indices in the original bytecode.

It also provides an `ExecutionThread` view type that allows manipulating
independent instruction pointers over the same instruction stream in
preparation for the `VM` needing to explore multiple threads of
execution.