Implement the `InstructionStream` and Parser (#13)
This type represents a sequence of EVM `Opcode`s, and maintains a 1:1
correspondence with the byte indices in the original bytecode.

It also provides an `ExecutionThread` view type that allows manipulating
independent instruction pointers over the same instruction stream in
preparation for the `VM` needing to explore multiple threads of
execution.

This adds tests for all the key scenarios of disassembling bytecode into
richer structures. It ensures that we can correctly roundtrip between
the `InstructionStream` and the external bytecode, as well as take those
instructions in both a byte array and a hex string.