Implement the `Storage` container
This is a version of the EVM's persistent storage designed to operate on
symbolic data. It maps from symbolic expressions (representing slots in
memory) to symbolic expressions (representing the values stored at those
locations).

This commit also introduces comprehensive provenance information
tracking to the `SymbolicValue` representation used during execution.
This ensures that the `Unifier` has access to the most information
possible about the potential type of a given storage location.