Implement `SymbolicValue` and the VM's `Stack` (#15)
This type represents the execution tree of operations done on symbolic
values at symbolic execution time. It is these trees that provide
evidence as to the type of a variable.

The stack is specialised for being the stack of the symbolic virtual
machine, and asserts length properties while also providing useful
operations for implementing the various `Opcode`s.