Implement the `Memory` for the `VM` (#16)
This commit implements the symbolic memory for the symbolic virtual
machine. It is capable of mapping from symbolic expressions to other
symbolic expressions, and thereby implementing `MSTORE*` and `MLOAD`
operations.

Note that it does _not_ account for overlapping writes, and if this
needs to be handled it should be handled at a higher level by
introducing synthetic expressions or decomposing expressions to check
for such reads. It is not feasible to do in general.