Refactor the structure of the errors (#26)
This commit refactors the way that the errors are structured to
better-categorise them into specific error types. It also introduces the
interface error type for the library.

In addition to this, it adds some execution tests on bytecode to the
`VM`, fixing some bugs encountered in the execution and error handling
logic therein.