Fix instruction tracking (#29)
Previously, instruction visitation was tracked at the level of the
symbolic virtual machine. This was incorrect, as it would cause
execution to bail out in valid sequences due to a different thread of
execution having already visited instructions.

Now, visitation is tracked at the level of the individual thread, and is
forked when the thread is forked, thus preventing a given sequence from
re-visiting something (and hence looping infinitely), but also allowing
individual sequences to visit all the instructions that they need.