Implement the end-to-end pipeline
For the first time the analyzer can be run from end to end, ingesting
bytecode and spitting out a (currently suboptimal) storage layout.

There are multiple cases that are currently not handled that will be
seen on real contracts, but this body of work is large as is. These
missing features will be improved in the future.