Fully implement unification (#36)
This approach performs all resolution in the space of the typing
inferences, using rules to collapse multiple inferences into one where
possible. It handles the conversion to the `AbiType` separately.

This PR also implements handling of CBOR metadata through the
recognition that it is not reachable in the bytecode. Hence, the
disassembler can just translate them to an opcode that errors if ever
executed.

This PR also starts using `clippy::pedantic` to improve code style and
hence includes a number of changes to please the linter. Overall, they
contribute to better and safer code.