Add some basic inference rules
For any immediately obvious type inference that can be made from each
node in the tree of Symbolic Values, we now have an inference rule that
associates that type information with those nodes.

None of the rules in this commit are complex, and they are not intended
to be so.