Introduce packed encodings
This commit introduces values in both the symbolic value representation
and type expression language that provide the analyzer with the means to
reason about packed encodings.

Note that it does not add support for discovering such encodings at this
stage, instead focusing on the supporting types needed to work with said
encodings.