Discover sub-word encodings (#41)
This commit adds a lifting pass that, where possible, discovers accesses
to portions of a word smaller than the entire word. This allows the
analyser to discover the size of types in many cases, but is also the
precursor work to supporting packed encodings into single storage slots.

In addition the codebase is now explicit about the byte ordering it is
constructed with, and maintains the "stored as little-endian" invariant
explicitly.

It also adds a test for packed encodings that is not yet properly
supported by the analyser, but is necessary for phase one of the
packed-encoding work.