Make execution tractable (#47)
The previous approach to exploring conditional branches meant that
contracts with nested conditionals would experience a combinatorial
explosion in the number of virtual machine threads waiting to execute.

This commit introduces a new approach that ensures that all instructions
in the bytecode are visited, but not all paths to all of those
instructions are executed. This drastically cuts the program state space
while still ensuring that we gather the required information.

This is done by imposing a (configurable) limit on the number of times a
given `JUMPDEST` can be jumped to by `JUMPI` instructions. Normal `JUMP`
instructions will always proceed as usual as they do not cause a
bifurcation in the state space.