Introduce the aux data field to Symbolic Value
This is determined statically in the type of the symbolic value,
allowing us to encode additional information in the values for different
stages of the analyzer.

In particular, it is designed to ensure that during execution we use
semantic equality for values, while during type checking we use
identity equality for values, where that identity equality is based on
the associated type variable.

This means that our type checker will no longer equate values that are
structurally equal but semantically different for typing purposes, and
hopefully give us more accurate results.