Fix spurious member in packed encodings
The logic for writing out packed encodings (of all kinds) was incorrect
at the last stage. It would sometimes insert a spurious member at offset
0 within the slot with type `Any` due to incorrect defaulting rules.
This is now fixed, and we output the packed encodings correctly.

We also now infer that the type of a storage load is the same as the
type of the slot being loaded from. This is correct in any well-formed
program and provides strictly more information to the type checker.