Introduce size tracking to the value stack (#57)
Some contracts exhibit pathological behaviour that ends up creating
values that are incredibly deep and grow as large as multiple GB for a
single value.

Construction of symbolic values now inspects the size of the value and
culls the value (replacing it with a bare `SVD::Value`) if it grows too
large.

As this behaviour only happens in very pathological cases, it only has a
minor impact on the accuracy of the analysis that is run. Nevertheless,
it ensures that operations can be run without allocations starting to
dominate computation time, and without excessive memory usage.

The default threshold of 500 for values is set at double the value
needed for all non-new tests to pass (~250). This ensures that we have
some headroom without making the threshold ineffective.