Introduce a watchdog mechanism
This provides a way for clients of the library to inject functionality
into the analyzer that tells the analyzer when it should stop its
analysis before completing.

It also provides a `LazyWatchdog` that has no ability to stop the
analyzer, and a `FlagWatchdog` that stops the analyzer based on changes
to an externally-accessible flag.