Fix a possibly diverging loop (#61)
Note that this commit introduces two new tests without specifying the
returned layouts. This is because these tests previously hung or
crashed, and now they do not. The layout specifications are to come.

It also fixes the constant folding operations to properly respect that
`a / 0 == 0` on the EVM. This was missing before.