Approximately halve memory usage (#66)
After the recent major refactor to fix a correctness bug in the
type-checker, the inference engine continued to hang onto the execution
result which it no longer needed. This is now no longer the case,
resulting in an approximate halving of maximum memory usage.

In addition, this also makes the maximum number of bytes that can be
operated on in a single VM memory operation configurable externally,
providing another point through which library clients can tune memory
usage.