Revive the inference rule tests (#84)
Due to changes in the operation of the inference state these would fail
in most cases, while still being able to compile. To that end, they were
disabled to benefit working speed, in the hope that they would be fixed
later.

This commit is that fix, which has worked out the best way to revive
these tests and have them working again.