Recognise the proxy slot pattern
This commit implements a feature for the analyzer to recognise the
commonly used proxy-slot pattern. It does this by a limited form of
constant folding for values that are likely to be hashes of encoded
strings.

This commit also improves the analyzer by causing it to constant fold
lengths in memory reads, thereby giving itself more accurate information
about the values actually being passed around in the program.