Share liberally in the analyzer
Prior to this commit, state forks would deep clone potentially quite
large value trees. This would create many duplicated values that were
then incorporated into further trees.

Now, rather than having unique ownership of each tree node using
`Box<T>`, the analyzer instead shares these as much as possible through
use of `Arc<T>`. This has resulted in significant improvements in memory
usage, and similarly significant improvements to execution time by
eliminating the need to perform many allocations.

In addition to this, we are also more careful about how we retain
values. When transforming `RSV` to `TCSV`, we would previously retain
the vector of all runtime values until all of those values had been
transformed. Now we drop runtime values eagerly wherever possible. This
has resulted in a reduction to peak residency.

As part of this work, some investigation was done as to the quadratic
traversal of value trees in the `InferenceState` during transformation.
While a variety of single-pass approaches were tried, the current
approach managed to outperform all of them. A comment has been added to
indicate this, and to be careful when changing it.