Refactor and document the code
This commit ensures that all implicit assumptions in the code are
documented explicitly, and refactors some structural elements to make it
clearer which portions of the code are responsible for which operations.

It also introduces `AbiType::Bits` to account for bytes-style usages
that do not have a size that is a multiple of `BYTE_SIZE`.